refactor: added io utils; moved check_zip
