+++ src/lib.rs
@@ -1,1 +1,2 @@
 pub mod cli;
+pub mod utils;
